move module
> Insert changelog entry or delete this section.
