+++ python_modules/dagster/dagster/_core/definitions/declarative_automation/operands/operands.py
@@ -1,7 +1,5 @@
 import datetime
-import os
-from collections.abc import Mapping, Sequence, Set
-from typing import TYPE_CHECKING, Optional
+from typing import Optional
 
 from dagster_shared.serdes import whitelist_for_serdes
 from dagster_shared.serdes.utils import SerializableTimeDelta
@@ -16,16 +14,10 @@ from dagster._core.definitions.declarative_automation.automation_context import 
 from dagster._core.definitions.declarative_automation.operands.subset_automation_condition import (
     SubsetAutomationCondition,
 )
-from dagster._core.definitions.events import AssetKeyPartitionKey
 from dagster._core.definitions.partitions.snap.snap import PartitionsSnap
 from dagster._core.definitions.partitions.subset.key_ranges import KeyRangesPartitionsSubset
-from dagster._core.loader import LoadingContext
 from dagster._record import record
 from dagster._utils.schedules import reverse_cron_string_iterator
-
-if TYPE_CHECKING:
-    from dagster._core.event_api import EventLogRecord
-    from dagster._core.storage.dagster_run import RunRecord
 
 
 @whitelist_for_serdes
@@ -183,250 +175,6 @@ class NewlyRequestedCondition(SubsetAutomationCondition):
 
     def compute_subset(self, context: AutomationContext) -> EntitySubset:
         return context.get_previous_requested_subset(context.key) or context.get_empty_subset()
-
-
-@whitelist_for_serdes
-@record
-class LatestRunExecutedWithRootTargetCondition(SubsetAutomationCondition):
-    @property
-    def name(self) -> str:
-        return "executed_with_root_target"
-
-    async def compute_subset(self, context: AutomationContext) -> EntitySubset:  # pyright: ignore[reportIncompatibleMethodOverride]
-        def _filter_fn(run_record: "RunRecord") -> bool:
-            if context.key == context.root_context.key:
-                # this happens when this is evaluated for a self-dependent asset. in these cases,
-                # it does not make sense to consider the asset as having been executed with itself
-                # as the partition key of the target is necessarily different than the partition
-                # key of the query key
-                return False
-            asset_selection = run_record.dagster_run.asset_selection or set()
-            check_selection = run_record.dagster_run.asset_check_selection or set()
-            return context.root_context.key in (asset_selection | check_selection)
-
-        return await context.asset_graph_view.compute_latest_run_matches_subset(
-            from_subset=context.candidate_subset, filter_fn=_filter_fn
-        )
-
-
-def _run_tag_filter_fn(
-    run_record: "RunRecord",
-    tag_keys: Optional[Set[str]],
-    tag_values: Optional[Mapping[str, str]],
-) -> bool:
-    if tag_keys and not all(key in run_record.dagster_run.tags for key in tag_keys):
-        return False
-    if tag_values and not all(
-        run_record.dagster_run.tags.get(key) == value for key, value in tag_values.items()
-    ):
-        return False
-    return True
-
-
-async def _get_run_records_from_materializations(
-    materializations: Sequence["EventLogRecord"],
-    context: LoadingContext,
-) -> Sequence["RunRecord"]:
-    from dagster._core.storage.dagster_run import RunRecord
-
-    run_ids = list({record.run_id for record in materializations if record.run_id})
-    if not run_ids:
-        return []
-
-    run_records = []
-    run_step = int(os.getenv("DAGSTER_ASSET_DAEMON_RUN_RECORDS_FROM_MATERIALIZATION_LIMIT", "1000"))
-    for i in range(0, len(run_ids), run_step):
-        chunk = run_ids[i : i + run_step]
-        chunk_records = await RunRecord.gen_many(
-            context,
-            chunk,
-        )
-        run_records.extend([record for record in chunk_records if record])
-
-    return run_records
-
-
-def _get_validated_partition_keys(
-    unvalidated_partition_keys: Set[Optional[str]],
-    context: AutomationContext,
-) -> Set[Optional[str]]:
-    partitions_def = context.asset_graph.get(context.key).partitions_def
-    if partitions_def is None:
-        return {
-            partition_key for partition_key in unvalidated_partition_keys if partition_key is None
-        }
-    else:
-        return {
-            partition_key
-            for partition_key in unvalidated_partition_keys
-            if partition_key is not None and partitions_def.has_partition_key(partition_key)
-        }
-
-
-@whitelist_for_serdes
-@record
-class LatestRunExecutedWithTagsCondition(SubsetAutomationCondition):
-    tag_keys: Optional[Set[str]] = None
-    tag_values: Optional[Mapping[str, str]] = None
-
-    @property
-    def name(self) -> str:
-        name = "executed_with_tags"
-        props = []
-        if self.tag_keys is not None:
-            tag_key_str = ",".join(sorted(self.tag_keys))
-            props.append(f"tag_keys={{{tag_key_str}}}")
-        if self.tag_values is not None:
-            tag_value_str = ",".join(
-                [f"{key}:{value}" for key, value in sorted(self.tag_values.items())]
-            )
-            props.append(f"tag_values={{{tag_value_str}}}")
-
-        if props:
-            name += f"({', '.join(props)})"
-        return name
-
-    async def compute_subset(self, context: AutomationContext) -> EntitySubset:  # pyright: ignore[reportIncompatibleMethodOverride]
-        return await context.asset_graph_view.compute_latest_run_matches_subset(
-            from_subset=context.candidate_subset,
-            filter_fn=lambda run_record: _run_tag_filter_fn(
-                run_record, self.tag_keys, self.tag_values
-            ),
-        )
-
-
-@whitelist_for_serdes
-@record
-class AnyNewMaterializationHasRunTagsCondition(SubsetAutomationCondition):
-    tag_keys: Optional[Set[str]] = None
-    tag_values: Optional[Mapping[str, str]] = None
-
-    @property
-    def name(self) -> str:
-        name = "any_new_materialization_has_run_tags"
-        props = []
-        if self.tag_keys is not None:
-            tag_key_str = ",".join(sorted(self.tag_keys))
-            props.append(f"tag_keys={{{tag_key_str}}}")
-        if self.tag_values is not None:
-            tag_value_str = ",".join(
-                [f"{key}:{value}" for key, value in sorted(self.tag_values.items())]
-            )
-            props.append(f"tag_values={{{tag_value_str}}}")
-
-        if props:
-            name += f"({', '.join(props)})"
-        return name
-
-    async def compute_subset(self, context: AutomationContext) -> EntitySubset:  # pyright: ignore[reportIncompatibleMethodOverride]
-        if (
-            not context.previous_temporal_context
-            or not context.previous_temporal_context.last_event_id
-        ):
-            return context.get_empty_subset()
-
-        new_materializations = context.asset_graph_view.get_inner_queryer_for_back_compat().get_asset_materializations_updated_after_cursor(
-            asset_key=context.key,
-            after_cursor=context.previous_temporal_context.last_event_id,
-        )
-        if not new_materializations:
-            return context.get_empty_subset()
-
-        run_records = await _get_run_records_from_materializations(
-            new_materializations,
-            context.asset_graph_view,
-        )
-
-        valid_run_ids: Set[str] = {
-            run_record.dagster_run.run_id
-            for run_record in run_records
-            if _run_tag_filter_fn(run_record, self.tag_keys, self.tag_values)
-        }
-
-        valid_partition_keys = set()
-
-        for materialization in new_materializations:
-            if materialization.run_id and (materialization.run_id in valid_run_ids):
-                valid_partition_keys.add(materialization.partition_key)
-
-        valid_partition_keys = _get_validated_partition_keys(valid_partition_keys, context)
-
-        return context.asset_graph_view.get_asset_subset_from_asset_partitions(
-            key=context.key,
-            asset_partitions={
-                AssetKeyPartitionKey(context.key, partition_key)
-                for partition_key in valid_partition_keys
-            },
-        )
-
-
-@whitelist_for_serdes
-@record
-class AllNewUpdatesHaveRunTagsCondition(SubsetAutomationCondition):
-    tag_keys: Optional[Set[str]] = None
-    tag_values: Optional[Mapping[str, str]] = None
-
-    @property
-    def name(self) -> str:
-        name = "all_new_updates_have_run_tags"
-        props = []
-        if self.tag_keys is not None:
-            tag_key_str = ",".join(sorted(self.tag_keys))
-            props.append(f"tag_keys={{{tag_key_str}}}")
-        if self.tag_values is not None:
-            tag_value_str = ",".join(
-                [f"{key}:{value}" for key, value in sorted(self.tag_values.items())]
-            )
-            props.append(f"tag_values={{{tag_value_str}}}")
-
-        if props:
-            name += f"({', '.join(props)})"
-        return name
-
-    async def compute_subset(self, context: AutomationContext) -> EntitySubset:  # pyright: ignore[reportIncompatibleMethodOverride]
-        if (
-            not context.previous_temporal_context
-            or not context.previous_temporal_context.last_event_id
-        ):
-            return context.get_empty_subset()
-
-        new_materializations = context.asset_graph_view.get_inner_queryer_for_back_compat().get_asset_materializations_updated_after_cursor(
-            asset_key=context.key,
-            after_cursor=context.previous_temporal_context.last_event_id,
-        )
-        if not new_materializations:
-            return context.get_empty_subset()
-
-        run_records = await _get_run_records_from_materializations(
-            new_materializations,
-            context.asset_graph_view,
-        )
-
-        invalid_run_ids = {
-            run_record.dagster_run.run_id
-            for run_record in run_records
-            if not _run_tag_filter_fn(run_record, self.tag_keys, self.tag_values)
-        }
-
-        materialized_partition_keys = set()
-        invalid_partition_keys = set()
-
-        for materialization in new_materializations:
-            materialized_partition_keys.add(materialization.partition_key)
-            if not materialization.run_id or (materialization.run_id in invalid_run_ids):
-                invalid_partition_keys.add(materialization.partition_key)
-
-        valid_partition_keys = _get_validated_partition_keys(
-            materialized_partition_keys - invalid_partition_keys, context
-        )
-
-        return context.asset_graph_view.get_asset_subset_from_asset_partitions(
-            key=context.key,
-            asset_partitions={
-                AssetKeyPartitionKey(context.key, partition_key)
-                for partition_key in valid_partition_keys
-            },
-        )
 
 
 @whitelist_for_serdes
