+++ python_modules/dagster/dagster_tests/declarative_automation_tests/automation_condition_tests/builtins/test_executed_with_tags_condition.py
@@ -234,9 +234,9 @@ def test_all_new_updates_have_run_tags_partitioned(small_fetch_batch_sizes) -> N
     assert result.total_requested == 0
 
 
-def test_any_new_materialization_has_run_tags(small_fetch_batch_sizes) -> None:
+def test_any_new_update_has_run_tags(small_fetch_batch_sizes) -> None:
     @dg.asset(
-        automation_condition=AutomationCondition.any_new_materialization_has_run_tags(
+        automation_condition=AutomationCondition.any_new_update_has_run_tags(
             tag_values={"target_tag": "a"}
         )
     )
@@ -277,9 +277,9 @@ def test_any_new_materialization_has_run_tags(small_fetch_batch_sizes) -> None:
     assert result.total_requested == 0
 
 
-def test_any_new_materialization_has_run_tags_partitioned(small_fetch_batch_sizes) -> None:
+def test_any_new_update_has_run_tags_partitioned(small_fetch_batch_sizes) -> None:
     @dg.asset(
-        automation_condition=AutomationCondition.any_new_materialization_has_run_tags(
+        automation_condition=AutomationCondition.any_new_update_has_run_tags(
             tag_values={"target_tag": "a"}
         ),
         partitions_def=dg.StaticPartitionsDefinition(["a", "b", "c"]),
